+++ src/main/java/ca/ulaval/glo4003/service/user/limit/MoneyAmountLimitDto.java
@@ -1,12 +1,13 @@
 package ca.ulaval.glo4003.service.user.limit;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class MoneyAmountLimitDto extends LimitDto {
 
-  public final double moneyAmount;
+  public final BigDecimal moneyAmount;
 
-  public MoneyAmountLimitDto(LocalDateTime from, LocalDateTime to, double moneyAmount) {
+  public MoneyAmountLimitDto(LocalDateTime from, LocalDateTime to, BigDecimal moneyAmount) {
     super(from, to);
     this.moneyAmount = moneyAmount;
   }
