+++ src/main/java/ca/ulaval/glo4003/ws/api/user/dto/ApiMoneyAmountLimitDto.java
@@ -1,6 +1,7 @@
 package ca.ulaval.glo4003.ws.api.user.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Schema(
@@ -11,9 +12,9 @@ public class ApiMoneyAmountLimitDto extends ApiLimitDto {
   @Schema(
       description = "The maximal money amount the user can spend in a single transaction."
   )
-  public final double maximalMoneySpent;
+  public final BigDecimal maximalMoneySpent;
 
-  public ApiMoneyAmountLimitDto(LocalDateTime start, LocalDateTime end, double maximalMoneySpent) {
+  public ApiMoneyAmountLimitDto(LocalDateTime start, LocalDateTime end, BigDecimal maximalMoneySpent) {
     super(start, end);
     this.maximalMoneySpent = maximalMoneySpent;
   }
